feat: export datacollection local storae
