+++ packages/react/src/lib/exports.ts
@@ -8,6 +8,7 @@ export * from "./dnd/types"
 export * from "./emojis"
 export { experimentalComponent as experimental } from "./experimental"
 export { PrivacyModeProvider, usePrivacyMode } from "./privacyMode"
+export * from "./providers/datacollection/exports"
 export { F0EventCatcherProvider } from "./providers/events"
 export { buildTranslations } from "./providers/i18n"
 export type { TranslationsType } from "./providers/i18n"
