+++ packages/react/src/lib/providers/datacollection/dataCollectionLocalStorageHandler.ts
@@ -1,7 +1,4 @@
-import {
-  DataCollectionStorage,
-  DataCollectionStorageHandler,
-} from "./DataCollectionStorageProvider"
+import { DataCollectionStorage, DataCollectionStorageHandler } from "./types"
 
 const getKey = (keyName: string) => `datacollection-${keyName}`
 export const dataCollectionLocalStorageHandler: DataCollectionStorageHandler = {
