+++ packages/react/src/lib/providers/datacollection/exports.ts
@@ -1,3 +1,2 @@
 export * from "./dataCollectionLocalStorageHandler"
-export * from "./DataCollectionStorageProvider"
 export * from "./types"
