+++ packages/react/src/lib/providers/f0/f0-provider.tsx
@@ -16,10 +16,8 @@ import { LinkContextValue, LinkProvider } from "../../linkHandler"
 import { PrivacyModeProvider } from "../../privacyMode"
 import { cn } from "../../utils"
 import { XRayProvider } from "../../xray"
-import {
-  DataCollectionStorageHandler,
-  DataCollectionStorageProvider,
-} from "../datacollection/DataCollectionStorageProvider"
+import { DataCollectionStorageProvider } from "../datacollection/DataCollectionStorageProvider"
+import { DataCollectionStorageHandler } from "../datacollection/types"
 
 interface LayoutProps {
   fullScreen?: boolean
